+++ wordcloud_mapper/__init__.py
@@ -3,6 +3,7 @@
 from .wordcloud_map import wordcloud_map
 from .load_companies import load_companies
 from .resize_map import resize_map
+from .calc_tfidf import calc_tfidf
 
 __author__ = """Gabriel da Silva Zech"""
 __email__ = 'g.dev@posteo.net'