+++ HISTORY.rst
@@ -14,3 +14,4 @@ History
 * Allow users to change the parameters when downloading NUTS shapefiles from Eurostat's API in wordcloud_map().
 * Allow users to change the sharpness of the regional border lines by channging the DPI value used when creating the masks.
 * Add wordcloud colour function based on rank of words.
+* Add new function ``calc_tfidf()`` to calculate TF-IDF score of each word in each region in a dataframe.